+++ conan/tools/cmake/utils.py
@@ -10,38 +10,6 @@ def is_multi_configuration(generator):
     return "Visual" in generator or "Xcode" in generator
 
 
-def architecture_flag(settings):
-    """
-    returns flags specific to the target architecture and compiler
-    """
-    compiler = settings.get_safe("compiler")
-    compiler_base = settings.get_safe("compiler.base")
-    arch = settings.get_safe("arch")
-    the_os = settings.get_safe("os")
-    if not compiler or not arch:
-        return ""
-
-    if str(compiler) in ['gcc', 'apple-clang', 'clang', 'sun-cc']:
-        if str(arch) in ['x86_64', 'sparcv9', 's390x']:
-            return '-m64'
-        elif str(arch) in ['x86', 'sparc']:
-            return '-m32'
-        elif str(arch) in ['s390']:
-            return '-m31'
-        elif str(the_os) == 'AIX':
-            if str(arch) in ['ppc32']:
-                return '-maix32'
-            elif str(arch) in ['ppc64']:
-                return '-maix64'
-    elif str(compiler) == "intel":
-        # https://software.intel.com/en-us/cpp-compiler-developer-guide-and-reference-m32-m64-qm32-qm64
-        if str(arch) == "x86":
-            return "/Qm32" if str(compiler_base) == "Visual Studio" else "-m32"
-        elif str(arch) == "x86_64":
-            return "/Qm64" if str(compiler_base) == "Visual Studio" else "-m64"
-    return ""
-
-
 def get_generator(conanfile):
     # Returns the name of the generator to be used by CMake
     if "CONAN_CMAKE_GENERATOR" in os.environ:
